Added Tempo functionality
A clock (in seconds) is running at the upper right side of the screen. The devices vibrate once in a second when there is less than 5s left. The character dies when time's out.

Currently level 1 has 45 seconds and each new level has -5 seconds.

Levels 3 and 5 are not completely connected.